+++ CHANGELOG.md
@@ -11,9 +11,11 @@
 ## Unreleased - in development
 ### Added
 - Added the new Tool `WfsSearch`.
+- New parameter "searchResultOrder" for ranking category of searching result
 
 ### Changed
 - Moved SensorThingsMqtt and SensorThingsHttp to /src/utils, complete refactoring of SensorThingsMqtt, there are no changes in handling SensorThingsMqtt.
+- Moved convertArrayOfObjectsToCsv to /src/utils/convertJsonToCsv.js with refactoring, removed convertArrayOfObjectsToCsv Event from Radio.
 
 ### Deprecated